+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-
+### Changed
+- riki137/stream-ipc to ^1.0 stable
+- Enhanced orchestrator resiliency by removing completed tasks after stream failures and logging IPC timeouts.
 
 ## [1.0.0-beta9] - 2025-08-14
 ### Added